+++ android/app/src/main/java/io/coursepick/coursepick/data/preferences/PreferencesDataStore.kt
@@ -6,6 +6,7 @@ import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.stringPreferencesKey
 import io.coursepick.coursepick.di.PreferencesDataStore
+import io.coursepick.coursepick.domain.preferences.RouteFinder
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
