+++ android/app/src/main/java/io/coursepick/coursepick/domain/preferences/PreferencesRepository.kt
@@ -1,6 +1,5 @@
 package io.coursepick.coursepick.domain.preferences
 
-import io.coursepick.coursepick.data.preferences.RouteFinder
 import kotlinx.coroutines.flow.Flow
 
 interface PreferencesRepository {
