+++ android/app/src/main/java/io/coursepick/coursepick/presentation/course/RouteFinderApplication.kt
@@ -4,8 +4,8 @@ import android.content.Intent
 import androidx.annotation.StringRes
 import androidx.core.net.toUri
 import io.coursepick.coursepick.R
-import io.coursepick.coursepick.data.preferences.RouteFinder
 import io.coursepick.coursepick.domain.course.Coordinate
+import io.coursepick.coursepick.domain.preferences.RouteFinder
 import kotlin.math.ln
 import kotlin.math.tan
