+++ android/app/src/main/java/io/coursepick/coursepick/presentation/preferences/PreferencesViewModel.kt
@@ -3,8 +3,8 @@ package io.coursepick.coursepick.presentation.preferences
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import io.coursepick.coursepick.data.preferences.RouteFinder
 import io.coursepick.coursepick.domain.preferences.PreferencesRepository
+import io.coursepick.coursepick.domain.preferences.RouteFinder
 import jakarta.inject.Inject
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
